Extract CoreData helper methods into a helper file
